refactor(paint): 重构外观设置页面，合并主题选择组件
1.  删除冗余的ColorThemeCard组件，将主题选择功能整合进AppearanceCard
2.  调整多个输入控件的统一宽度为240px优化界面对齐
3.  新增分组标题拆分外观设置为主题和显示两个区块
4.  修复多语言文件新增分组翻译键值
5.  调整设置页面的通用和外观区块顺序
