+++ src/language/ja-JP.json
@@ -649,6 +649,8 @@
     "max_log_lines_desc": "サーバーごとの最大ログ行数。超過分は自動削除。デフォルト5000",
     "appearance": "外観",
     "appearance_desc": "背景や視覚効果のカスタマイズ",
+    "group_theme": "テーマ",
+    "group_effect": "表示",
     "theme": "テーマモード",
     "theme_desc": "アプリケーションのテーマを選択。「システムに従う」でOSの設定に連動",
     "font_size": "テキストサイズ",
