+++ src/language/ko-KR.json
@@ -649,6 +649,8 @@
     "max_log_lines_desc": "서버당 최대 로그 보관 줄 수입니다. 초과 시 오래된 로그부터 자동 삭제됩니다. 기본값 5000",
     "appearance": "외관",
     "appearance_desc": "소프트웨어 배경 및 시각 효과 사용자 정의",
+    "group_theme": "테마",
+    "group_effect": "표시",
     "theme": "테마 모드",
     "theme_desc": "애플리케이션 테마를 선택합니다. '시스템 설정'은 시스템의 다크/라이트 모드를 자동으로 따릅니다",
     "font_size": "텍스트 크기",
