+++ src/language/ru-RU.json
@@ -649,6 +649,8 @@
     "max_log_lines_desc": "Максимальное количество строк лога для одного сервера. Старые автоматически удаляются. По умолч. 5000",
     "appearance": "Внешний вид",
     "appearance_desc": "Настройка фона и визуальных эффектов",
+    "group_theme": "Тема",
+    "group_effect": "Отображение",
     "theme": "Тема",
     "theme_desc": "Выберите тему оформления",
     "font_size": "Размер текста",
