+++ src/language/zh-TW.json
@@ -650,6 +650,8 @@
     "max_log_lines_desc": "單個伺服器最多保留的日誌行數，超出後自動清除舊日誌。預設 5000",
     "appearance": "外觀",
     "appearance_desc": "自訂軟體背景和視覺效果",
+    "group_theme": "主題",
+    "group_effect": "顯示",
     "theme": "主題模式",
     "theme_desc": "選擇應用程式的主題外觀，\"跟隨系統\"會自動匹配系統的深色/淺色模式",
     "font_size": "文字大小",
